Add 'ValueOnly' data type to AddressingModeOutput enum
The `Fetched` data type needs to store both the fetched value and the
address it has been fetched from. In order for this to happen
additional field is needed, so that the addressing modes which only
return value (e.g `Implied` and `Immediate`) can be used properly. This
field is called `ValueOnly`.

(All tests are updated to check the new return types and are
 successfully passing.)