Merge v0.01 m6502 implementation dev branch (#3)
* Add 'ValueOnly' data type to AddressingModeOutput enum

The `Fetched` data type needs to store both the fetched value and the
address it has been fetched from. In order for this to happen
additional field is needed, so that the addressing modes which only
return value (e.g `Implied` and `Immediate`) can be used properly. This
field is called `ValueOnly`.

(All tests are updated to check the new return types and are
 successfully passing.)

* Add ADC, AND, ASL instructions

* Add branching instructions (BCC, BCS, BMI, BNE, BPL, BVC, BVS)

Each one of them is implemented using a helping function which wraps the
actual branching logic. The specific branch instruction checks its
predicate and if is correct a branch will be made.

* Add BRK, BIT and all "Clear * flag" instructions

* Add all "load" and "store" instructions

The LDA, LDX, LDY are considered "load" instructions and STA, STX, STY
are considered - "store".

* Add "increment" instructions

These include INX, INY, INC.

* Add compare instructions: CMP, CPY, CPX

* Add return instructions: RTI, RTS

* Add monitors for cpu internals and memory

* Demonstrate monitor usage with a fibonacci sequence program

* Finish implementation of the instructions associated with legal opcodes

Implementation to the following instructions has been included:

- the "decrement" instructions: **DEC**, **DEY**, **DEX**
- **EOR**, **ORA**
- **JMP**, **JSR**
- **LSR** + additional change to **ASL**
- the "push" instructions: **PHA**, **PHP**
- the "pull" instructions: **PLA**, **PLP**
- the "rotate" instructions: **ROL**, **ROR**
- **RTS**
- the "flag set" instructions: **SEI**, **SED**, **SEC**
- **NOP**
- the "substract" instruction: **SBC**
- the transfer instructions: **TAX**, **TAY**, **TSX**, **TXA**,
    **TXS**, **TYA**

An additional changes have been made to the following:
- **ASL** - Cleanup

To all instructions has been a brief description.

Some of the instructions have associated unit tests.

* Add gcd example